requirements.txt
google
aiohttp
python-dotenv
requests
discord.py
flask
